Добавил реализацию forwardRef
